Repair systemd reload notifications
In order for the RELOAD and the subsequent READY notifications to be
correctly processed, the RELOAD notification must be accompanied with a
MONOTONIC_USEC one.
Unfortunately, the sd-notify crate does not seem to support this, and I
have not yet looked into implementing it.
Hence this patch is comprised of:
- replacement of sd-notify with systemd for the daemon component to
  support the MONOTONIC_USEC notification
- adding of the MONOTONIC_USEC notification to repair the RELOAD =>
  READY notifications
- adding of STATUS notifications for both easier debugging of whether
  notifications arrive with systemd and easier user facing feedback
  about the current daemon status
- removal of the artificial delay, which is no longer beneficial
- disabling of unset_environment, as the preservation of $NOTIFY_SOCKET
  is crucial

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>