Remove systemd notify-reload
systemd notify-reload has been nothing but problems for us. Since
it was introduced, systemd broke it once by forcing a new parameter
to be added to messages, and to further that the parameter they
added was documented so poorly we couldn't get it implemented
correctly.

Remove notify-reload, and return to classic - and reliable - HUP
signals.