Avoid openssl for md4
Since we are in the process of dropping openssl, we are no longer
limited by it. In this case, the issue is that openssl demands the
legacy provider be enabled for FreeIPA sync due to it's weak
cryptography.

The issue however is that with the changes to the scratch container
base, we no longer have the legacy module, and working to include it
isn't worth the effort since we can just swap to rust crypto instead.