Set kid manually to prevent divergence
Due to a quirk of how compact-jwt worked, the Key ID (kid) that it
would use for signing may be different to the kid that Kanidm was
using to reference the key. This was due to a change in the kid
format, that wasn't fully handled.

Now on key load, the correct kid can be set by Kanidm so that
all signatures have a kid that matches what Kanidm believes the
kid to be.