Move shadow processing out of task event loop (#3631)
As part of the recent unixd changes, we moved shadow processing to the privileged tasks daemon. In that move the handling of those files was moved into the event loop of the task handler itself, which could cause the daemon to become unresponsive for short or long periods, which then caused unixd to fail to make home directories.

This moves the shadow processing to a separate task, and only once it's ready the state flips and notifies the connected unixd caller.