Add user facing SCIM pagination / sorting (#3725)
This adds the user facing parts of SCIM pagination and sorting.
Currently it's done in a very "simple" way, with sorting and
truncating the result vector. This won't be the long term
approach however, we need to add a proper set of listing and
sorting indexes to the backend. But having some working tests
is always good to show where we want to go with this.

Part of the problem is that sorting and pagination indexes are
hard to do, so I need to think more about the best way to add
them :)