The client tool config had the wrong port specified
The client tool config had port 8443 on localhost specified in the uri. This is the internal docker port from the evaluation example server.toml, not a port on the host running docker. The evaluation example sets the port to 443 on localhost, the standard https port. If login is attempted on port 8443 on localhost, the connection will be refused.

Therefore either a redundant port specification needs to be made, or the port specification needs to be entirely removed, and https://localhost will connect to the correct port already.