Reduce memory usage on unixd
This introduces two more changes to reduce memory consumption with
kanidm-unixd.

The first is that there is a potential issue with mimalloc where it
can cause the RSS to be elevated. As unixd is not as performance
sensitive as the main server, we limit mimalloc only to linux for
the main server daemon instead, as glibc malloc has a number of
issues when under high demand workloads.

The second was that due to pbkdf2 benchmarking at server startup,
the memory reports from dhat were reporting those blocks as high
usage. Since pbkdf is NOT a format we use, we now statically
allocate it's round numbers in the unlikely case we do ever use
it. The important algorithm remains argon2id for our applications.