Resolve replication show-cert issue
The replication show cert command relies on the replication loop
being operational to display the cert. When we changed to rustls
the issue is that empty cert stores are not valid for their PKI
verifier, which caused the replication accept loop to never start
meaning that show-cert would never complete.

This resolves that by handling the empty cert store case specially
and rejecting all incomming connections when the store is empty.