Add json codec wrapper for unix integration (#3789)
* Add json codec wrapper for unix integration

Previously our IPC was simple JSON over UnixStreams. This has worked
well enough for a long time, but as we have expanded the capabilities
of tasks and the resolver, there is now much more traffic between the
two.

This can lead to a situation where when writing or decoding frames,
the buffer if it contains multiple frames will only read the first
frame and then clear() itself. This causes message loss resulting
in odd behaviour and dropped events.

To resolve this, we cleanup codec handling and make a generic
JsonCodec wrapper. This wrapper now prepends a length to each message
so that we can properly remove the bytes of the frame without
affecting subsequent frames that may be present.

* Update comment